move map files into map directory
