+++ client/src/components/map/MapContainer.js
@@ -3,8 +3,8 @@ import MapView from 'react-native-maps';
 import { View, Dimensions } from 'react-native';
 import { Actions } from 'react-native-router-flux';
 import geolib from 'geolib';
-import getSpots from '../lib/getSpots';
-import getLatLong from '../lib/getLatLong';
+import getSpots from '../../lib/getSpots';
+import getLatLong from '../../lib/getLatLong';
 import AddPhotoIcon from './AddPhotoIcon';
 import LocateSelfIcon from './LocateSelfIcon';
 import LensIcon from './LensIcon';
